Removing aggregate keyword
Fixes #246

Removing aggregate as keyword in actual VSS specification,
but leaving it as allowed syntax when defining a branch.
(I.e. vss-tools shall still accept usage of it)

The intention to leave it is as it might be reused in e.g.
the future layer concept, i.e. an instantiation of VSS might
specify that a particular branch shall be handled as an aggregate.